tests/wpt/web-platform-tests/xhr/abort-event-listeners.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() should not reset event listeners</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-abort()-method" data-tested-assertations="following-sibling::ol/li[6] following-sibling::ol/li[7]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            test = function() {}
        client.onreadystatechange = test
        client.open("GET", "resources/well-formed.xml")
        client.send(null)
        client.abort()
        assert_equals(client.onreadystatechange, test)
      })
      test.done()
    </script>
  </body>
</html>
